feat: 自动添加按钮 (#48)
* chore: bump deps

* chore: don't force yarn to help local dev

* feat: 自动添加按钮

* fix: revert pkg mgr changes.

---------
